# core/src/main/resources/swagger_spec.json
{
  "swagger" : "2.0",
  "info" : {
    "version" : "1.2.0",
    "title" : "Klaw Core",
    "contact" : {
      "email" : "info@klaw-project.io"
    }
  },
  "basePath" : "/",
  "paths" : {
...
    "/addNewTeam" : {
      "post" : {
        "operationId" : "addNewTeam",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/TeamModel"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/ApiResponse"
            }
          }
        }
      }
    },
    "/addNewUser" : {
      "post" : {
        "operationId" : "addNewUser",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/UserInfoModel"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/ApiResponse"
            }
          }
        }
      }
    },
...
    "/chPwd" : {
      "post" : {
        "operationId" : "changePwd",
        "parameters" : [ {
          "name" : "changePwd",
          "in" : "query",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/ApiResponse"
            }
          }
        }
      }
    },
...
    "/deleteTeamRequest" : {
      "post" : {
        "operationId" : "deleteTeam",
        "parameters" : [ {
          "name" : "teamId",
          "in" : "query",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/ApiResponse"
            }
          }
        }
      }
    },
...
    "/deleteUserRequest" : {
      "post" : {
        "operationId" : "deleteUser",
        "parameters" : [ {
          "name" : "userId",
          "in" : "query",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/ApiResponse"
            }
          }
        }
      }
    },
...
    "/execNewUserRequestApprove" : {
      "post" : {
        "operationId" : "approveNewUserRequests",
        "parameters" : [ {
          "name" : "username",
          "in" : "query",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/ApiResponse"
            }
          }
        }
      }
    },
    "/execNewUserRequestDecline" : {
      "post" : {
        "operationId" : "declineNewUserRequests",
        "parameters" : [ {
          "name" : "username",
          "in" : "query",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/ApiResponse"
            }
          }
        }
      }
    },
...
    "/getActivationInfo" : {
      "get" : {
        "operationId" : "getActivationInfo",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "userActivationId",
          "in" : "query",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "object",
              "additionalProperties" : {
                "type" : "string"
              }
            }
          }
        }
      }
    },
...
    "/getTeamDetails" : {
      "get" : {
        "operationId" : "getTeamDetails",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "teamId",
          "in" : "query",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        }, {
          "name" : "tenantName",
          "in" : "query",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/TeamModel"
            }
          }
        }
      }
    },
...
    "/getUserDetails" : {
      "get" : {
        "operationId" : "getUserDetails",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "userId",
          "in" : "query",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/UserInfoModel"
            }
          }
        }
      }
    },
    "/getUserInfoFromRegistrationId" : {
      "get" : {
        "operationId" : "getRegistrationInfoFromId",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "userRegistrationId",
          "in" : "query",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/RegisterUserInfoModel"
            }
          }
        }
      }
    },
...
    "/registerUser" : {
      "post" : {
        "operationId" : "registerUser",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/RegisterUserInfoModel"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/ApiResponse"
            }
          }
        }
      }
    },
    "/registerUserSaas" : {
      "post" : {
        "operationId" : "registerUserSaas",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/RegisterSaasUserInfoModel"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/ApiResponse"
            }
          }
        }
      }
    },
...
    "/resetPassword" : {
      "post" : {
        "operationId" : "resetPassword",
        "parameters" : [ {
          "name" : "username",
          "in" : "query",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "object",
              "additionalProperties" : {
                "type" : "string"
              }
            }
          }
        }
      }
    },
...
    "/showUserList" : {
      "get" : {
        "operationId" : "showUsers",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "teamName",
          "in" : "query",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "pageNo",
          "in" : "query",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "searchUserParam",
          "in" : "query",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/UserInfoModel"
              }
            }
          }
        }
      }
    },
...
    "/updateProfile" : {
      "post" : {
        "operationId" : "updateProfile",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/UserInfoModel"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/ApiResponse"
            }
          }
        }
      }
    },
...
    "/updateTeam" : {
      "post" : {
        "operationId" : "updateTeam",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/TeamModel"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/ApiResponse"
            }
          }
        }
      }
    },
    "/updateUser" : {
      "post" : {
        "operationId" : "updateUser",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/UserInfoModel"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/ApiResponse"
            }
          }
        }
      }
    },
...
  },
  "definitions" : {
...
    "ApiResponse" : {
      "type" : "object",
      "properties" : {
        "status" : {
          "type" : "string",
          "enum" : [ "CONTINUE", "SWITCHING_PROTOCOLS", "PROCESSING", "CHECKPOINT", "OK", "CREATED", "ACCEPTED", "NON_AUTHORITATIVE_INFORMATION", "NO_CONTENT", "RESET_CONTENT", "PARTIAL_CONTENT", "MULTI_STATUS", "ALREADY_REPORTED", "IM_USED", "MULTIPLE_CHOICES", "MOVED_PERMANENTLY", "FOUND", "MOVED_TEMPORARILY", "SEE_OTHER", "NOT_MODIFIED", "USE_PROXY", "TEMPORARY_REDIRECT", "PERMANENT_REDIRECT", "BAD_REQUEST", "UNAUTHORIZED", "PAYMENT_REQUIRED", "FORBIDDEN", "NOT_FOUND", "METHOD_NOT_ALLOWED", "NOT_ACCEPTABLE", "PROXY_AUTHENTICATION_REQUIRED", "REQUEST_TIMEOUT", "CONFLICT", "GONE", "LENGTH_REQUIRED", "PRECONDITION_FAILED", "PAYLOAD_TOO_LARGE", "REQUEST_ENTITY_TOO_LARGE", "URI_TOO_LONG", "REQUEST_URI_TOO_LONG", "UNSUPPORTED_MEDIA_TYPE", "REQUESTED_RANGE_NOT_SATISFIABLE", "EXPECTATION_FAILED", "I_AM_A_TEAPOT", "INSUFFICIENT_SPACE_ON_RESOURCE", "METHOD_FAILURE", "DESTINATION_LOCKED", "UNPROCESSABLE_ENTITY", "LOCKED", "FAILED_DEPENDENCY", "UPGRADE_REQUIRED", "PRECONDITION_REQUIRED", "TOO_MANY_REQUESTS", "REQUEST_HEADER_FIELDS_TOO_LARGE", "UNAVAILABLE_FOR_LEGAL_REASONS", "INTERNAL_SERVER_ERROR", "NOT_IMPLEMENTED", "BAD_GATEWAY", "SERVICE_UNAVAILABLE", "GATEWAY_TIMEOUT", "HTTP_VERSION_NOT_SUPPORTED", "VARIANT_ALSO_NEGOTIATES", "INSUFFICIENT_STORAGE", "LOOP_DETECTED", "BANDWIDTH_LIMIT_EXCEEDED", "NOT_EXTENDED", "NETWORK_AUTHENTICATION_REQUIRED" ]
        },
        "timestamp" : {
          "type" : "string",
          "format" : "date-time"
        },
        "message" : {
          "type" : "string"
        },
        "debugMessage" : {
          "type" : "string"
        },
        "result" : {
          "type" : "string"
        },
        "data" : {
          "type" : "object"
        }
      }
    },
...
    "RegisterSaasUserInfoModel" : {
      "type" : "object",
      "properties" : {
        "fullname" : {
          "type" : "string"
        },
        "mailid" : {
          "type" : "string"
        },
        "recaptchaStr" : {
          "type" : "string"
        },
        "status" : {
          "type" : "string"
        },
        "registeredTime" : {
          "type" : "string",
          "format" : "date-time"
        },
        "approver" : {
          "type" : "string"
        },
        "registrationId" : {
          "type" : "string"
        },
        "tenantName" : {
          "type" : "string"
        }
      }
    },
    "RegisterUserInfoModel" : {
      "type" : "object",
      "properties" : {
        "username" : {
          "type" : "string"
        },
        "pwd" : {
          "type" : "string"
        },
        "team" : {
          "type" : "string"
        },
        "teamId" : {
          "type" : "integer",
          "format" : "int32"
        },
        "role" : {
          "type" : "string"
        },
        "fullname" : {
          "type" : "string"
        },
        "mailid" : {
          "type" : "string"
        },
        "status" : {
          "type" : "string"
        },
        "registeredTime" : {
          "type" : "string",
          "format" : "date-time"
        },
        "approver" : {
          "type" : "string"
        },
        "registrationId" : {
          "type" : "string"
        },
        "tenantId" : {
          "type" : "integer",
          "format" : "int32"
        },
        "tenantName" : {
          "type" : "string"
        }
      }
    },
...
    "TeamModel" : {
      "type" : "object",
      "properties" : {
        "teamname" : {
          "type" : "string"
        },
        "teammail" : {
          "type" : "string"
        },
        "teamphone" : {
          "type" : "string"
        },
        "contactperson" : {
          "type" : "string"
        },
        "tenantId" : {
          "type" : "integer",
          "format" : "int32"
        },
        "teamId" : {
          "type" : "integer",
          "format" : "int32"
        },
        "app" : {
          "type" : "string"
        },
        "showDeleteTeam" : {
          "type" : "boolean"
        },
        "tenantName" : {
          "type" : "string"
        },
        "envList" : {
          "type" : "array",
          "items" : {
            "type" : "string"
          }
        }
      }
    },
...
    "UserInfoModel" : {
      "type" : "object",
      "properties" : {
        "username" : {
          "type" : "string"
        },
        "userPassword" : {
          "type" : "string"
        },
        "team" : {
          "type" : "string"
        },
        "teamId" : {
          "type" : "integer",
          "format" : "int32"
        },
        "role" : {
          "type" : "string"
        },
        "fullname" : {
          "type" : "string"
        },
        "mailid" : {
          "type" : "string"
        },
        "tenantName" : {
          "type" : "string"
        },
        "tenantId" : {
          "type" : "integer",
          "format" : "int32"
        },
        "totalNoPages" : {
          "type" : "string"
        },
        "allPageNos" : {
          "type" : "array",
          "items" : {
            "type" : "string"
          }
        }
      }
    },
...
  }
}